requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
openai==1.59.5
